+++ src/main/java/com/yapp/artie/global/deprecated/UserJpaEntityLoader.java
@@ -1,7 +1,7 @@
 package com.yapp.artie.global.deprecated;
 
 import com.yapp.artie.domain.user.adapter.out.persistence.UserJpaEntity;
-import com.yapp.artie.domain.user.application.port.exception.UserNotFoundException;
+import com.yapp.artie.domain.user.application.service.UserNotFoundException;
 import java.util.Optional;
 import javax.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
